Update README.md
Attached in the SQL tabs are some examples of work i completed for a company that helped them identify customers and invoices to help identify ways to increase production. 
In the Alteryx tab it incudes a snapshot of an automation i created that helped a stakeholder automate a report that they previously had to manually find and document that took up about 20 hours of their time each week.
In the Dual Access, Tableau and Simplify2 tab are some snapshots of examples i created for a powerpoint for UHC to show how to make the data reportings more clear, simpler, and easy to read at first glance.